+++ Sources/KeyboardKit/Gestures/Gestures+ScrollViewGestureButton.swift
@@ -50,7 +50,7 @@ public extension Gestures {
            - endAction: The action to trigger when a button gesture ends, by default `nil`.
            - label: The button label.
          */
-        init(
+        public init(
             isPressed: Binding<Bool>? = nil,
             pressAction: Action? = nil,
             releaseInsideAction: Action? = nil,